# ARROW-8658: [C++][Dataset] Implement subtree pruning for FileSystemDataset