--- a/cpp/src/arrow/dataset/expression_benchmark.cc
+++ b/cpp/src/arrow/dataset/expression_benchmark.cc
@@ -15,8 +15,6 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include <iostream>
-
 #include "benchmark/benchmark.h"
 
 #include "arrow/compute/cast.h"