--- a/cpp/src/arrow/dataset/file_base.cc
+++ b/cpp/src/arrow/dataset/file_base.cc
@@ -24,6 +24,7 @@
 #include <vector>
 
 #include "arrow/dataset/dataset_internal.h"
+#include "arrow/dataset/expression_internal.h"
 #include "arrow/dataset/scanner.h"
 #include "arrow/dataset/scanner_internal.h"
 #include "arrow/filesystem/filesystem.h"
@@ -36,6 +37,7 @@
 #include "arrow/util/make_unique.h"
 #include "arrow/util/map.h"
 #include "arrow/util/mutex.h"
+#include "arrow/util/optional.h"
 #include "arrow/util/string.h"
 #include "arrow/util/task_group.h"
 
@@ -82,30 +84,23 @@ Result<ScanTaskIterator> FileFragment::Scan(std::shared_ptr<ScanOptions> options
   return format_->ScanFile(std::move(options), std::move(context), self);
 }
 
-FileSystemDataset::FileSystemDataset(std::shared_ptr<Schema> schema,
-                                     Expression root_partition,
-                                     std::shared_ptr<FileFormat> format,
-                                     std::shared_ptr<fs::FileSystem> filesystem,
-                                     std::vector<std::shared_ptr<FileFragment>> fragments)
-    : Dataset(std::move(schema), std::move(root_partition)),
-      format_(std::move(format)),
-      filesystem_(std::move(filesystem)),
-      fragments_(std::move(fragments)) {}
-
 Result<std::shared_ptr<FileSystemDataset>> FileSystemDataset::Make(
     std::shared_ptr<Schema> schema, Expression root_partition,
     std::shared_ptr<FileFormat> format, std::shared_ptr<fs::FileSystem> filesystem,
     std::vector<std::shared_ptr<FileFragment>> fragments) {
-  return std::shared_ptr<FileSystemDataset>(new FileSystemDataset(
-      std::move(schema), std::move(root_partition), std::move(format),
-      std::move(filesystem), std::move(fragments)));
+  std::shared_ptr<FileSystemDataset> out(
+      new FileSystemDataset(std::move(schema), std::move(root_partition)));
+  out->format_ = std::move(format);
+  out->filesystem_ = std::move(filesystem);
+  out->fragments_ = std::move(fragments);
+  out->SetupSubtreePruning();
+  return out;
 }
 
 Result<std::shared_ptr<Dataset>> FileSystemDataset::ReplaceSchema(
     std::shared_ptr<Schema> schema) const {
   RETURN_NOT_OK(CheckProjectable(*schema_, *schema));
-  return std::shared_ptr<Dataset>(new FileSystemDataset(
-      std::move(schema), partition_expression_, format_, filesystem_, fragments_));
+  return Make(std::move(schema), partition_expression_, format_, filesystem_, fragments_);
 }
 
 std::vector<std::string> FileSystemDataset::files() const {
@@ -137,16 +132,174 @@ std::string FileSystemDataset::ToString() const {
   return repr;
 }
 
-Result<FragmentIterator> FileSystemDataset::GetFragmentsImpl(Expression predicate) {
-  FragmentVector fragments;
+namespace {
 
-  for (const auto& fragment : fragments_) {
-    ARROW_ASSIGN_OR_RAISE(
-        auto simplified,
-        SimplifyWithGuarantee(predicate, fragment->partition_expression()));
-    if (simplified.IsSatisfiable()) {
-      fragments.push_back(fragment);
+// Helper class for efficiently detecting subtrees given fragment partition expressions.
+// Partition expressions are broken into conjunction members and each member dictionary
+// encoded to impose a sortable ordering. In addition, subtrees are generated which span
+// groups of fragments and nested subtrees. After encoding each fragment is guaranteed to
+// be a descendant of at least one subtree. For example, given fragments in a
+// HivePartitioning with paths:
+//
+//   /num=0/al=eh/dat.par
+//   /num=0/al=be/dat.par
+//   /num=1/al=eh/dat.par
+//   /num=1/al=be/dat.par
+//
+// The following subtrees will be introduced:
+//
+//   /num=0/
+//   /num=0/al=eh/
+//   /num=0/al=eh/dat.par
+//   /num=0/al=be/
+//   /num=0/al=be/dat.par
+//   /num=1/
+//   /num=1/al=eh/
+//   /num=1/al=eh/dat.par
+//   /num=1/al=be/
+//   /num=1/al=be/dat.par
+struct SubtreeImpl {
+  using expression_code = char32_t;
+  using expression_codes = std::basic_string<expression_code>;
+
+  std::unordered_map<Expression, expression_code, Expression::Hash> expr_to_code_;
+  std::vector<Expression> code_to_expr_;
+  std::unordered_set<expression_codes> subtree_exprs_;
+
+  // An encoded fragment (if fragment_index is set) or subtree.
+  struct Encoded {
+    util::optional<int> fragment_index;
+    expression_codes partition_expression;
+  };
+
+  expression_code GetOrInsert(const Expression& expr) {
+    auto next_code = static_cast<int>(expr_to_code_.size());
+    auto it_success = expr_to_code_.emplace(expr, next_code);
+
+    if (it_success.second) {
+      code_to_expr_.push_back(expr);
     }
+    return it_success.first->second;
+  }
+
+  void EncodeConjunctionMembers(const Expression& expr, expression_codes* codes) {
+    if (auto call = expr.call()) {
+      if (call->function_name == "and_kleene") {
+        // expr is a conjunction, encode its arguments
+        EncodeConjunctionMembers(call->arguments[0], codes);
+        EncodeConjunctionMembers(call->arguments[1], codes);
+        return;
+      }
+    }
+    // expr is not a conjunction, encode it whole
+    codes->push_back(GetOrInsert(expr));
+  }
+
+  Expression GetSubtreeExpression(const Encoded& encoded_subtree) {
+    // Filters will already be simplified by all of a subtree's ancestors, so
+    // we only need to simplify the filter by the trailing conjunction member
+    // of each subtree.
+    return code_to_expr_[encoded_subtree.partition_expression.back()];
+  }
+
+  void GenerateSubtrees(expression_codes partition_expression,
+                        std::vector<Encoded>* encoded) {
+    while (!partition_expression.empty()) {
+      if (subtree_exprs_.insert(partition_expression).second) {
+        Encoded encoded_subtree{/*fragment_index=*/util::nullopt, partition_expression};
+        encoded->push_back(std::move(encoded_subtree));
+      }
+      partition_expression.resize(partition_expression.size() - 1);
+    }
+  }
+
+  void EncodeOneFragment(int fragment_index, const Fragment& fragment,
+                         std::vector<Encoded>* encoded) {
+    Encoded encoded_fragment{fragment_index, {}};
+
+    EncodeConjunctionMembers(fragment.partition_expression(),
+                             &encoded_fragment.partition_expression);
+
+    GenerateSubtrees(encoded_fragment.partition_expression, encoded);
+
+    encoded->push_back(std::move(encoded_fragment));
+  }
+
+  template <typename Fragments>
+  std::vector<Encoded> EncodeFragments(const Fragments& fragments) {
+    std::vector<Encoded> encoded;
+    for (size_t i = 0; i < fragments.size(); ++i) {
+      EncodeOneFragment(static_cast<int>(i), *fragments[i], &encoded);
+    }
+    return encoded;
+  }
+};
+
+}  // namespace
+
+void FileSystemDataset::SetupSubtreePruning() {
+  SubtreeImpl impl;
+
+  auto encoded = impl.EncodeFragments(fragments_);
+
+  std::sort(encoded.begin(), encoded.end(),
+            [](const SubtreeImpl::Encoded& l, const SubtreeImpl::Encoded& r) {
+              return l.partition_expression < r.partition_expression;
+            });
+
+  for (const auto& e : encoded) {
+    if (e.fragment_index) {
+      fragments_and_subtrees_.emplace_back(*e.fragment_index);
+    } else {
+      fragments_and_subtrees_.emplace_back(impl.GetSubtreeExpression(e));
+    }
+  }
+
+  forest_ = fs::Forest(static_cast<int>(encoded.size()), [&](int l, int r) {
+    if (encoded[l].fragment_index) {
+      // Fragment: not an ancestor.
+      return false;
+    }
+
+    const auto& ancestor = encoded[l].partition_expression;
+    const auto& descendant = encoded[r].partition_expression;
+
+    if (descendant.size() >= ancestor.size()) {
+      return std::equal(ancestor.begin(), ancestor.end(), descendant.begin());
+    }
+    return false;
+  });
+}
+
+Result<FragmentIterator> FileSystemDataset::GetFragmentsImpl(Expression predicate) {
+  std::vector<int> fragment_indices;
+
+  std::vector<Expression> predicates{predicate};
+  RETURN_NOT_OK(forest_.Visit(
+      [&](fs::Forest::Ref ref) -> Result<bool> {
+        if (auto fragment_index = util::get_if<int>(&fragments_and_subtrees_[ref.i])) {
+          fragment_indices.push_back(*fragment_index);
+          return false;
+        }
+
+        const auto& subtree_expr = util::get<Expression>(fragments_and_subtrees_[ref.i]);
+        ARROW_ASSIGN_OR_RAISE(auto simplified,
+                              SimplifyWithGuarantee(predicates.back(), subtree_expr));
+
+        if (!simplified.IsSatisfiable()) {
+          return false;
+        }
+
+        predicates.push_back(std::move(simplified));
+        return true;
+      },
+      [&](fs::Forest::Ref ref) { predicates.pop_back(); }));
+
+  std::sort(fragment_indices.begin(), fragment_indices.end());
+
+  FragmentVector fragments;
+  for (int i : fragment_indices) {
+    fragments.push_back(fragments_[i]);
   }
 
   return MakeVectorIterator(std::move(fragments));