… cpp/src/arrow/dataset/CMakeLists.txt
… cpp/src/arrow/dataset/file_benchmark.cc
… cpp/src/arrow/dataset/forest.h
… cpp/src/arrow/dataset/forest_internal.h
… cpp/src/arrow/filesystem/CMakeLists.txt
… cpp/src/arrow/filesystem/path_forest.cc
… cpp/src/arrow/filesystem/path_forest.h
… cpp/src/arrow/filesystem/path_forest_test.cc