TODO:
- [ ] Add benchmarks to be sure this provides an advantage
- [ ] Move Forest to dataset_internal.h or so, since it is not used anywhere else
- [ ] Unit test SubtreeImpl, add more comments