@lidavidm 2021-03-10
On a quick benchmark, this is a two orders of magnitude speedup! Though I'm also going to test some other scenarios + add unit tests.

Before:

```
---------------------------------------------------------------
Benchmark                     Time             CPU   Iterations
---------------------------------------------------------------
GetAllFragments        83434318 ns     83433416 ns           17
GetFilteredFragments  502654017 ns    502651704 ns            3
```

After:

```
---------------------------------------------------------------
Benchmark                     Time             CPU   Iterations
---------------------------------------------------------------
GetAllFragments        72942933 ns     72941992 ns           19
GetFilteredFragments    5630666 ns      5630548 ns          250
```