@lidavidm 2021-03-11
Before:
```
---------------------------------------------------------------------------
Benchmark                                 Time             CPU   Iterations
---------------------------------------------------------------------------
GetAllFragments                    77918155 ns     77914414 ns           18
GetFilteredFragments/single_dir   504830505 ns    504788164 ns            3
GetFilteredFragments/single_file 1135530755 ns   1135494963 ns            1
GetFilteredFragments/range        509878962 ns    509855374 ns            3
```

After:
```
---------------------------------------------------------------------------
Benchmark                                 Time             CPU   Iterations
---------------------------------------------------------------------------
GetAllFragments                    67112654 ns     67106684 ns           21
GetFilteredFragments/single_dir     5564513 ns      5564443 ns          246
GetFilteredFragments/single_file   12560697 ns     12560479 ns          111
GetFilteredFragments/range         71451702 ns     71449949 ns           19
```