@lidavidm 2021-03-12
MacOS tests are fixed now that the sorting on subtrees is fully defined.